+++ src/main/kotlin/kr/hhplus/be/server/api/support/filter/QueueTokenValidInterceptor.kt
@@ -2,8 +2,8 @@ package kr.hhplus.be.server.api.support.filter
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import kr.hhplus.be.server.domain.common.error.CoreApiException
-import kr.hhplus.be.server.domain.common.error.ErrorType
+import kr.hhplus.be.server.domain.support.error.CoreException
+import kr.hhplus.be.server.domain.support.error.ErrorType
 import kr.hhplus.be.server.domain.token.TokenService
 import org.springframework.stereotype.Component
 import org.springframework.web.servlet.HandlerInterceptor
@@ -15,10 +15,10 @@ class QueueTokenValidInterceptor(
 ) : HandlerInterceptor {
 
     override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
-        val token = request.getHeader("QUEUE_TOKEN") ?: throw CoreApiException(ErrorType.BAD_REQUEST)
+        val token = request.getHeader("QUEUE_TOKEN") ?: throw CoreException(ErrorType.BAD_REQUEST)
         val isActive = tokenService.isActive(UUID.fromString(token))
         if (!isActive) {
-            throw CoreApiException(ErrorType.BAD_REQUEST)
+            throw CoreException(ErrorType.BAD_REQUEST)
         }
         return true
     }
